src/python/Text_extraction_video_v2/requirements.txt
pytesseract
wand
opencv-contrib-python
Pillow==9.3.0